Adding a function to search for old, inconsistent values
Adding a FindOldEntriesFunction that can be invoked like this:

execute function --id=FindOldEntriesFunction   --arguments=example-region,1

The arguments are a region name and a number of minutes. It will find any
entries that have a last modified time older than that number of minutes. It
will then check to see if the primary has an entry from that key. If the entry
is missing from the primary, it will be reported.